web_app/requirements.txt
Django==4.2.7
Pillow==10.1.0
ultralytics==8.0.214
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
segmentation_models_pytorch==0.3.3
